UdyamBharat/requirements.txt
Flask==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
Flask-Login==0.5.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
requests==2.31.0
Pillow==10.0.0
python-multipart==0.0.6
bcrypt==3.2.0
PyJWT==2.8.0
assemblyai==0.15.0
elevenlabs==0.2.24
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0 
google-generativeai==0.3.0 